+++ lib/tabs/legal/cookie_policy.dart
@@ -1,6 +1,8 @@
 import "package:flutter/material.dart";
 import "package:get/get.dart";
+import "package:x_containers/x_containers.dart";
 
+import "../../utils/utils.dart";
 import "legal_tab.dart";
 
 /// A tab to display the cookie policy of the website
@@ -17,6 +19,35 @@ class CookiePolicy extends StatelessWidget {
     return TabLegal(
       title: "Cookie policy".tr,
       text: "cookie_policy.md".tr,
+      appendix: XContainer(
+        padding: EdgeInsets.symmetric(
+          horizontal: XLayout.paddingL,
+          vertical: XLayout.paddingM,
+        ),
+        child: Row(
+          children: [
+            Text(
+              "Delete and disable cookies".tr,
+            ),
+            const Expanded(child: SizedBox()),
+            TextButton(
+              style: PresetStyles.secondaryButtonStyle(context),
+              onPressed: () {
+                settings.cookies.value = false;
+                XSnackbar.text(
+                  maxWidth: maxSnackbarLength,
+                  title: "Done.",
+                  content: "Your cookies were deleted and disabled.",
+                ).show(context);
+              },
+              child: Padding(
+                padding: EdgeInsets.all(XLayout.paddingS),
+                child: Text("Disable".tr),
+              ),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }